+++ CHANGELOG.md
@@ -12,6 +12,9 @@ Note to self: See the bottom of this file for the release template text.
  - Add `//sshift`  (_selection shift_) - WorldEdit cuboid manipulator replacements implemented by @VorTechnix.
  - Use [luacheck](https://github.com/mpeterv/luacheck) to find and fix a large number of bugs and other issues
  - Multiple commands: Allow using quotes (`"thing"`, `'thing'`) to quote values when splitting
+ - `//layers`: Add optional slope constraint (inspired by [WorldPainter](https://worldpainter.net/))
+ - `//bonemeal`: Add optional node list constraint
+ - `//walls`: Add optional thickness argument
 
 
 ## v1.12: The selection tools update (26th June 2021)